Moved components to same folder. Added styling to be used in new PR
