+++ frontend/src/components/Settings.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import NavBar from "../components/NavBar";
+import NavBar from "./NavBar";
 
 class Settings extends Component {
     render() {
